Support for libgnokii. Now we can connect a GSM phone to the server for sending SMS messages.
This should probably be modified in the future,
to use some sms-daemon, because the way it
implemented now, uw_process is the only app
allowed access to the phone.

Use the file upgrade-for-sms.mysql to upgrade 
your database for the new SMS fields.

Don't forget to give the upwatch user access to
the serial port, and the uucp lock directory.